ADD : adding my code for assignment 2
Using Multiple Linear Regression(MLR) to obtain the governing ordinary differential equation for the evolution of population of two creatures,the data of which is provided in the means of a csv file. We train and validate the model using 3 different splits, displaying the MSE(Mean Squared Error) when compared to the actual data.