ADD: Adding Midterm Project
This project revolves around enhancing the quality of transformer current data
sampled every 5 minutes, sourced from a solar power generation site. The dataset
likely contains fluctuations, anomalies, and missing values due to various factors such
as weather conditions, equipment malfunctions, or measurement errors. Our
objective is to identify 'good', 'bad', and 'missing' days within this dataset and improve
the quality of the data by addressing inconsistencies and filling in missing values.
Leveraging machine learning techniques, we aim to develop a model using data from
the identified 'good' days to learn patterns and relationships, which will then be
validated and utilized to enhance the quality of data from 'bad' and 'missing' days.
Through this process, we intend to provide a more reliable dataset for analysis and
decision-making in the context of solar power generation.